ocr-img2txt/ocr-img2txt/requirements.txt
gradio
pytesseract
Pillow
easyocr
